# speed up a couple of really slow unittests